First_review_20210204
Added a whole directory to review the main blocks of code of the tool.

There are only 2 templates so far, run and example. And there may be comments with "#!" that will be erased.

The schemas and configuration (both yaml files) will be updated along the templates to satisfy what information we want to provide. For now they are simply mock examples following the XML examples from ENA.