EE-822_First_review_feedback
Changed a few things following Anna's guidance, and mostly changed the structure of the files. Specially the star2xml.py and validateXML.py.

Until there's a propper documentation, if you want to check the new code with a few mock examples, use:
$ ./star2xml.py --schema-file 'xml_schema.yaml' --configuration-file 'input_configuration.yaml' 'run' 'final_templates/xlsx_templates/Filled_examples/run_template.xlsx' 'output_folder/run.xml' --verbose

$ ./star2xml.py --schema-file 'xml_schema.yaml' --configuration-file 'input_configuration.yaml' 'sample' 'final_templates/xlsx_templates/Filled_examples/Sample.xlsx' 'output_folder/sample.xml' --verbose

$ ./validateXML.py "sample,run" "output_folder/sample.xml,output_folder/run.xml" --schemas-dir "xsd_schemas/" --schema-file "xml_schema.yaml" --verbose --download_xsd

For now the configuration and schema YAML files are not fully completed (new fields and nodes will be added, etc.).